+++ apps/blog/app/layout.tsx
@@ -2,11 +2,11 @@ import "@xionwcfm/token/style";
 import "@xionwcfm/xds/style";
 import "./globals.css";
 
-import { AnalyticsProvider } from "@repo/analytics";
-import { env } from "@repo/env";
 import { Suspense } from "@suspensive/react";
 import { Toaster } from "@xionwcfm/xds/toast";
 import type { Metadata } from "next";
+import { AnalyticsProvider } from "~/shared/packages/analytics";
+import { env } from "~/shared/packages/env";
 import { Pretendard } from "../src/shared/assets/fonts";
 import { AUTHOR_NAME, BASE_SITE_DESCRIPTION, BASE_SITE_TITLE, BASE_SITE_URL } from "../src/shared/constants";
 import { createMetadata } from "../src/shared/utils/external/createMetaData";
@@ -31,12 +31,10 @@ export default function RootLayout({
   return (
     <html lang="ko">
       <body className={Pretendard.className}>
-        <Suspense>
-          <AnalyticsProvider>
-            {children}
-            <Toaster visibleToasts={1} />
-          </AnalyticsProvider>
-        </Suspense>
+        <AnalyticsProvider>
+          <Suspense>{children}</Suspense>
+          <Toaster visibleToasts={1} />
+        </AnalyticsProvider>
       </body>
     </html>
   );
