--- apps/blog/app/loading.tsx
@@ -1,3 +1,0 @@
-export default function Page() {
-  return <div></div>;
-}
