+++ apps/blog/package.json
@@ -22,13 +22,13 @@
     "@next/third-parties": "15.3.2",
     "@radix-ui/react-icons": "1.3.2",
     "@radix-ui/react-separator": "1.1.7",
-    "@repo/analytics": "workspace:*",
     "@repo/database": "workspace:*",
     "@repo/env": "workspace:*",
     "@repo/router": "workspace:*",
     "@supabase/ssr": "0.6.1",
     "@supabase/supabase-js": "2.49.6",
     "@suspensive/react": "3.2.1",
+    "@t3-oss/env-core": "0.13.4",
     "@types/mdx": "2.0.13",
     "@xionwcfm/react": "0.1.3",
     "@xionwcfm/token": "0.2.0",
