--- packages/analytics/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "@repo/analytics",
-  "version": "0.0.0",
-  "private": true,
-  "main": "./src/index.tsx",
-  "exports": {
-    ".": "./src/index.tsx",
-    "./use-feature-flag": "./src/useFeatureFlag.ts",
-    "./ab-test": "./src/AbTest.tsx",
-    "./posthog/client": "./src/posthog/client.tsx",
-    "./posthog/server": "./src/posthog/server.tsx"
-  },
-  "dependencies": {
-    "@next/third-parties": "15.3.2",
-    "@repo/env": "workspace:*",
-    "server-only": "0.0.1",
-    "react": "catalog:react18",
-    "react-dom": "catalog:react18"
-  },
-  "devDependencies": {
-    "@repo/typescript-config": "workspace:*",
-    "@types/node": "22.15.19",
-    "@types/react": "catalog:react18",
-    "@types/react-dom": "catalog:react18"
-  },
-  "peerDependencies": {
-    "next": ">=15",
-    "react": "catalog:react18"
-  }
-}
