--- packages/analytics/src/useFeatureFlag.ts
@@ -1,6 +1,0 @@
-import { useFeatureFlagEnabled } from "posthog-js/react";
-
-export const useFeatureFlag = (featureFlag: string) => {
-  const isEnabled = useFeatureFlagEnabled(featureFlag);
-  return isEnabled;
-};
